chore(provider): move `CanonStateNotification` to `reth-provider-types`
