+++ crates/storage/provider-types/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "reth-provider-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "Reth storage provider types"
+
+[lints]
+workspace = true
+
+[dependencies]
+# reth
+reth-execution-types.workspace = true
+reth-primitives.workspace = true
+
