+++ crates/storage/provider/Cargo.toml
@@ -30,6 +30,7 @@ reth-trie = { workspace = true, features = ["metrics"] }
 reth-nippy-jar.workspace = true
 reth-codecs.workspace = true
 reth-evm.workspace = true
+reth-provider-types.workspace = true
 
 # ethereum
 alloy-rpc-types-engine.workspace = true
